cambios

+++ src/Comentario.cpp
@@ -5,12 +5,14 @@
         this->texto = "";
         this->fecha = DTFecha(0,0,0);
         this->id = "";
+        this->producto = NULL;
     }
 
-    Comentario :: Comentario(string texto, DTFecha fecha, string id) {
+    Comentario :: Comentario(string texto, DTFecha fecha, string id, Producto * prod) {
         this->texto = texto;
         this->fecha = fecha;
         this->id = id;
+        this->producto = prod;
     }
 
     string Comentario :: getTexto() {
@@ -46,8 +48,12 @@
     }
 
     Comentario :: ~Comentario() {
-        this->producto->eliminarComentario(this->id);
-        this->producto = NULL;
+        if (this->producto != NULL) {
+            cout << "Desvinculando Comentario de Producto..." << endl;
+            this->producto->eliminarComentario(this->id);
+            this->producto = NULL;
+        }
+        //falta desvincular al usuario
         if (!this->respuestas.empty()) {
             for (Comentario * resp : this->respuestas)
                 delete resp;
